test(settings): pin retarget behavior for stores with no subscriber
Red without the follow-up fix: a store that warmed its cache via
value(for:) and has no values(for:) subscriber gets no watcher-driven
invalidation (drains spawn on first subscribe), so after the config
symlink is retargeted, reads keep serving the old target's root and a
set() writes that stale root over the new target — dropping its keys.

Also split the symlink suite into JSONConfigStoreSymlinkTests.swift
with shared scaffolding in JSONConfigStoreTestSupport.swift to keep
every test file under the 500-line budget.
